requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pypdf==5.9.0
python-docx==1.2.0
python-pptx==1.0.2
sentence-transformers==5.0.0
langchain==0.3.27
langchain-text-splitters==0.3.9
pymongo==4.13.2
motor==3.7.1
pinecone==7.3.0
google-generativeai==0.3.2
python-dotenv==1.1.1
httpx==0.28.1
torch==2.7.1
transformers==4.54.0
numpy==2.1.2
pillow==11.3.0
PyPDF2==3.0.1